# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VS Furniture API",
        "description": "API documentation for VS Furniture Laravel Backend",
        "contact": {
            "name": "VS Furniture",
            "email": "admin@vsfurniture.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders (admin only)",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully"
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new order",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "customer_phone",
                                    "items"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "customer_phone": {
                                        "type": "string"
                                    },
                                    "customer_address": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    }
                }
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get single order (admin only)",
                "operationId": "3a2d78e76e9fc3f06bb879fb7d48b40d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{order}/status": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status (admin only)",
                "operationId": "5b0239241bfbedd17e8148c3caf08679",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/stats/overview": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order statistics (admin only)",
                "operationId": "c374b7b23239c4ca9df8f483eebe43d8",
                "responses": {
                    "200": {
                        "description": "Order statistics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}